refactor(rpc): extract flat plan encoder
Move TxExecuteReadPlan flat payload encoding out of lineairdb_rpc.cc into flat_plan_encode.{cc,hh}.
Register flat_plan_encode.cc in the server build.
